Wrote an entire exercise in assembly
I have managed to slay my former beasts, with function callers and
loops all implemented in assembly, this puny exercise added to my
understanding of the stack and some new register names; rdi, rsi, rdx,
rcx, r8 and r9 which are 1st to 6th arguments to a function, in which
the rest are stored in the stack. Also, I wrote all assembly externally
and linked to the main program, this is an amazing step.